Add automation to retry delayed notifications
I observed over the past week that notifications would
occasionally be triggered in the Chrome API but would not be
visible in the Windows Notification bar.

I toggled with some notifications settings over the week but that
did not help. I hypothesized that the notifications were blocked
and needed to be 're-triggered' in some way, so I utilized the
'updateAlarm' Chrome API endpoint to trigger a 'refresh' of the
push.

Looks like it works for now! I added a new, more frequent timer
that updates 'stuck' alarms that I found in the 'getAllAlarms'
Chrome API endpoint, eventually triggering them to be sent to the
Windows Notifications API.