Pull alarm handlers out of sub-function and register them at top level
According to documentation for the chrome background pages found here
https://developer.chrome.com/extensions/background_pages, if listeners
or handlers are not set at the top level, they will not be registered
correctly.

My intuition tells me that Javascript does some dank process isolation
and listeners are probably reaped based on the lifetime of the parent
process. If a handler is not set at the top level, then the lifetime
of the handler is attached to the sub-function, which is ephemeral.

From my observations, this change has caused the alarms to
consistently be shown again.